begin semonar 7

+++ homeworks/homework-07/work3.py
@@ -11,15 +11,18 @@ import time
 
 
 def delete_old_files(src_dir, num_days):
-    src_dir = os.path.abspath(src_dir)
-    print(f"Start delete older files from {src_dir}")
-    last_time = time.time() - (num_days * 24 * 60 * 60)
-    for cur_dir, dirs, files in os.walk(src_dir):
-        for fn in files:
-            full_name = os.path.join(cur_dir, fn)
-            if os.path.getmtime(full_name) < last_time:
-                print(f'  - delete file {os.path.join(cur_dir, fn)}')
-                os.remove(full_name)
+    if os.path.isdir(src_dir):
+        src_dir = os.path.abspath(src_dir)
+        print(f"Start delete older files from {src_dir}")
+        last_time = time.time() - (num_days * 24 * 60 * 60)
+        for cur_dir, dirs, files in os.walk(src_dir):
+            for fn in files:
+                full_name = os.path.join(cur_dir, fn)
+                if os.path.getmtime(full_name) < last_time:
+                    print(f'  - delete file {os.path.join(cur_dir, fn)}')
+                    os.remove(full_name)
+    else:
+        print(f"Error: first parameter must be directory!")
 
 
 if __name__ == '__main__':
